# src/app/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PSL Studio API",
    "description": "REST API for PSL Studio — a fashion product lifecycle management platform covering products, samples, events, campaigns, loans, media placements, and guests.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from POST /api/auth/login or POST /api/auth/register. Pass in the Authorization header as: Bearer <token>"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        },
        "required": ["error"]
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true }
        }
      },
...
      "Campaign": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "deletedAt": { "type": "string", "format": "date-time", "nullable": true },
          "profileId": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "type": {
            "type": "string",
            "enum": ["DIGITAL", "PRINT", "OOH", "SOCIAL", "INFLUENCER", "OTHER"]
          },
          "status": {
            "type": "string",
            "enum": ["DRAFT", "ACTIVE", "PAUSED", "COMPLETED", "CANCELLED"]
          },
          "startAt": { "type": "string", "format": "date-time", "nullable": true },
          "endAt": { "type": "string", "format": "date-time", "nullable": true },
          "budget": { "type": "number", "nullable": true },
          "currency": { "type": "string", "nullable": true, "example": "EUR" },
          "eventId": { "type": "string", "nullable": true }
        }
      },
...
      "CampaignPatch": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 200 },
          "description": { "type": "string", "maxLength": 2000, "nullable": true },
          "type": {
            "type": "string",
            "enum": ["DIGITAL", "PRINT", "OOH", "SOCIAL", "INFLUENCER", "OTHER"]
          },
          "status": {
            "type": "string",
            "enum": ["DRAFT", "ACTIVE", "PAUSED", "COMPLETED", "CANCELLED"]
          },
          "startAt": { "type": "string", "format": "date-time", "nullable": true },
          "endAt": { "type": "string", "format": "date-time", "nullable": true },
          "budget": { "type": "number", "minimum": 0, "exclusiveMinimum": true, "nullable": true },
          "currency": { "type": "string", "minLength": 3, "maxLength": 3 },
          "eventId": { "type": "string", "nullable": true }
        }
      },
...
      "EventGuest": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "eventId": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "email": { "type": "string", "format": "email", "nullable": true },
          "phone": { "type": "string", "nullable": true },
          "instagram": { "type": "string", "nullable": true },
          "company": { "type": "string", "nullable": true },
          "title": { "type": "string", "nullable": true },
          "category": {
            "type": "string",
            "enum": ["VIP", "PRESS", "BUYER", "INFLUENCER", "INDUSTRY", "GUEST"]
          },
          "invitedBy": { "type": "string", "nullable": true },
          "rsvpStatus": {
            "type": "string",
            "enum": ["INVITED", "CONFIRMED", "DECLINED", "WAITLIST"]
          },
          "checkedIn": { "type": "boolean" },
          "checkedInAt": { "type": "string", "format": "date-time", "nullable": true },
          "tableNumber": { "type": "string", "nullable": true },
          "seatNumber": { "type": "string", "nullable": true },
          "notes": { "type": "string", "nullable": true }
        }
      },
      "GuestCreate": {
        "type": "object",
        "required": ["firstName", "lastName"],
        "properties": {
          "firstName": { "type": "string", "minLength": 1, "maxLength": 100 },
          "lastName": { "type": "string", "minLength": 1, "maxLength": 100 },
          "email": { "type": "string", "format": "email", "nullable": true },
          "phone": { "type": "string", "maxLength": 50, "nullable": true },
          "instagram": { "type": "string", "maxLength": 100, "nullable": true },
          "company": { "type": "string", "maxLength": 200, "nullable": true },
          "title": { "type": "string", "maxLength": 200, "nullable": true },
          "category": {
            "type": "string",
            "enum": ["VIP", "PRESS", "BUYER", "INFLUENCER", "INDUSTRY", "GUEST"],
            "default": "GUEST"
          },
          "invitedBy": { "type": "string", "maxLength": 200, "nullable": true },
          "rsvpStatus": {
            "type": "string",
            "enum": ["INVITED", "CONFIRMED", "DECLINED", "WAITLIST"],
            "default": "INVITED"
          },
          "tableNumber": { "type": "string", "maxLength": 20, "nullable": true },
          "seatNumber": { "type": "string", "maxLength": 20, "nullable": true },
          "notes": { "type": "string", "maxLength": 500, "nullable": true }
        }
      },
...
      "CampaignProductBody": {
        "type": "object",
        "required": ["productId"],
        "properties": {
          "productId": { "type": "string" },
          "notes": { "type": "string", "nullable": true }
        }
      },
      "CampaignProductDelete": {
        "type": "object",
        "required": ["productId"],
        "properties": {
          "productId": { "type": "string" }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 8, "maxLength": 100 }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" }
        }
      },
...
    },
    "parameters": {
      "idParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Resource identifier (cuid)"
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid authentication token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Non authentifié" }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request body or parameters",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    }
  },
...
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new account",
        "operationId": "authRegister",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created successfully. A session cookie is set.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": {
            "description": "Email already in use",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Un compte avec cet email existe déjà" }
              }
            }
          },
          "429": {
            "description": "Too many requests (rate limited: 5 per 15-minute window per IP)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/api/events/{id}/guests": {
      "get": {
        "tags": ["Guests"],
        "summary": "List guests for an event",
        "operationId": "listEventGuests",
        "parameters": [
          { "$ref": "#/components/parameters/idParam" }
        ],
        "responses": {
          "200": {
            "description": "Array of guests, sorted by category then last name",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/EventGuest" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Guests"],
        "summary": "Add a guest to an event",
        "operationId": "createEventGuest",
        "parameters": [
          { "$ref": "#/components/parameters/idParam" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GuestCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Guest created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EventGuest" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/campaigns/{id}": {
      "get": {
        "tags": ["Campaigns"],
        "summary": "Get a campaign by ID (with products and event)",
        "operationId": "getCampaign",
        "parameters": [
          { "$ref": "#/components/parameters/idParam" }
        ],
        "responses": {
          "200": {
            "description": "Campaign detail with products and linked event",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Campaign" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Campaigns"],
        "summary": "Partially update a campaign",
        "operationId": "patchCampaign",
        "parameters": [
          { "$ref": "#/components/parameters/idParam" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CampaignPatch" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated campaign",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Campaign" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Campaigns"],
        "summary": "Soft-delete a campaign",
        "operationId": "deleteCampaign",
        "parameters": [
          { "$ref": "#/components/parameters/idParam" }
        ],
        "responses": {
          "200": {
            "description": "Campaign soft-deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Success" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/campaigns/{id}/products": {
      "post": {
        "tags": ["Campaigns"],
        "summary": "Add or update a product in a campaign (upsert)",
        "operationId": "addCampaignProduct",
        "parameters": [
          { "$ref": "#/components/parameters/idParam" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CampaignProductBody" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product added to campaign",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "campaignId": { "type": "string" },
                    "productId": { "type": "string" },
                    "addedAt": { "type": "string", "format": "date-time" },
                    "notes": { "type": "string", "nullable": true }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "delete": {
        "tags": ["Campaigns"],
        "summary": "Remove a product from a campaign",
        "operationId": "removeCampaignProduct",
        "parameters": [
          { "$ref": "#/components/parameters/idParam" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CampaignProductDelete" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product removed from campaign",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Success" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  }
}